Separate client code into its own packages everywhere
