+++ deprecated/fabric-networking-v0/src/main/resources/fabric.mod.json
@@ -20,7 +20,7 @@
       "net.fabricmc.fabric.impl.networking.v0.OldNetworkingHooks"
     ],
     "client": [
-      "net.fabricmc.fabric.impl.networking.v0.OldClientNetworkingHooks"
+      "net.fabricmc.fabric.impl.client.networking.v0.OldClientNetworkingHooks"
     ]
   },
   "depends": {
