+++ fabric-networking-api-v1/src/client/resources/fabric-networking-api-v1.client.mixins.json
@@ -1,13 +1,13 @@
 {
   "required": true,
-  "package": "net.fabricmc.fabric.mixin.networking",
+  "package": "net.fabricmc.fabric.mixin.networking.client",
   "compatibilityLevel": "JAVA_16",
   "client": [
     "accessor.ConnectScreenAccessor",
     "accessor.MinecraftClientAccessor",
-    "client.ClientLoginNetworkHandlerMixin",
-    "client.ClientPlayNetworkHandlerMixin",
-    "client.DisconnectedScreenMixin"
+    "ClientLoginNetworkHandlerMixin",
+    "ClientPlayNetworkHandlerMixin",
+    "DisconnectedScreenMixin"
   ],
   "injectors": {
     "defaultRequire": 1
